passports work!
Passports now spawn in backpacks on arrival to the city.

A lockbox of temporary passports spawns in the City Clerk's office.

Temporary Passports - If a player loses their passport (stolen, destroyed, just plain ol' lost), they can go to the City Clerk and request a new temporary passport. I suggest a fee for requesting one. To set your information, click the temporary passport in-hand and it will match your character's information to the "Citizenship Database" and automatically update its details.

Diplomatic Passport - A passport for diplomats! Functions similar to temporary passports since diplomats don't spawn at the beginning of the round. There is an additional option that allows the diplomat to set the colony they are representing.

Antag options - Passports can be emagged allowing their details to be edited. You can set the name, and citizenship to whatever you want. Be warned, this only works ONCE! Diplomatic passports are immune for the time being.