Returns Weapons Crates to normal + Raises Cargo Prices
Weapon crates returned to normal health, in exchange, the price is heightened into the thousands, as it should be.

You can now ship any item, as long as it has a price tag (the type that's compatible with the atm), it will give you money as long as you send it in a crate.

Money now returns its own worth. V cool.

All things in cargo get price rises.