Bounty System
This is a bounty system which gives people bounties for their department which is a source of income, with this, I will be re-enabling service charges so be prepared.

* Generates a new bounty per hour.
* Bounties and their progress are persistent.
* Bounties pay for both individuals and the department.

Hope this all works, tried to sieve through for bugs and fixed what I found. If there's any just buzz me through git issues.
*