Adds Jewelry Presses
* Adds jewelry presses

The purpose of this machine (which is a subtype of autolathe) is to make more demand on certain materials. It has a high capacity for precious metals, and thus is a huge drain per item. This will force items that are made through it, IE: Jewelry, medals, trophies, to have an extremely high value due to how expensive it is.