[Speedmerge] Prevents Inv-Tri Item Duplication Glitch
Urgent hotfix. The glitch secret's out. Need to prevent big bags or other laggy items from being duped (since saving can lag for huge objs)

This moves the item in the invtri before the saving starts (which should solve the issue) but also as an extra layer of security by making the inv-tri stop working if it's processing an item.