Adds Autodocs
Autodoc machines are exclusively found in Geminus City Hospital and cannot be ordered/used anywhere else. These are ported from Eris, but I've edited them so people can operate on other people with it if they pay themselves, not just the person inside - hopefully this means dead people on lowpop can be saved.

Money charged from this machine goes right to healthcare dept funds and can finally be a good source of income. Because of autodocs I have readded the medical requirement access to hospital-use wallmeds (except for ones in the lobby that require purchases) so that people can't get freebies if they break in.

Scan Cost: 200 (Scan has to be done in order for organs to be fixed)
Organ Damage Cost: 800
Embed Objects: 1000
Fractures: 1200
Blood Replenish: 1200
Open Wounds: 600
Toxin Cost: 600
Dialysis: 1000

Prices are set quite high and even extortionate, the intent of this is so they're used as an emergency (so people stop looting the hospital).

Defibs are placed in purchaseable wallmeds in the hospital so that once a person has recieved treatment from an autodoc, they can be revived. Hurrah!