Gun Nerf, Part 1: The Breakening
First in the series of curbing the America-tier level of guns out there.

Gives all guns a usage limit (it buffers and has variation per gun, this can be tweaked per gun and by model). This is persistent. Once a gun is broken, you can keep it for nostalgia purposes, but it's a dead gun jim.

Can probably add gun modification and tinkering later, but this is currently for the purposes of somewhat taming the gun saturation for now.