Adds new weather type: Earthquakes
We need more natural disasters. So here we are.

Earthquakes are a new weather type that can occur on Pollux, when they do occur, they can be low intensity or high intensity. Usually it is harmless but you should lay down on the ground because there is a chance you may get flung off your feet into something and stunned if you are standing.

It affects both indoors and outdoors so beware.