Adds some fixes to the hospital and the PSDI Building
Fixes the missing walls for the Hospital and the floor turf issue causing it to snow inside the Hospital.

Fixes the PDSI's Doors and adds some more things to flesh it out more.

Adds A fax machine and water dispenser to the PDSI office space.

Adds a private reading corner to PDSI so they can be book worms.