Additions and Fixes
Hopefully Fixes the Medical Cryo Pod Issue, Adds two other Cryo Pods, One next to the old one and another on the second floor in Intensive Care.

Adds an Office space for Police, Six desks.

Adds a second floor to the Government funded housing block, changes the hallways of lot 4-5 to be unsaveable.